requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.32.0,<2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0